Documenting the difference between an entity head and ancillary heads.
Some heads play a role in repo structure and those should not be confused
with actual permanamed entities.

It's useful to be able to see all the details of someone else's repo and see
where they have been and where they are headed, but data sharing generally
only happens on permanamed entity boundaries.